requirements.txt
blivet==3.4.0
Brlapi==0.6.7
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.0.4
configobj==5.0.6
cupshelpers==1.0
dasbus==1.2
dataclasses==0.8
dbus-python==1.2.4
decorator==4.2.1
file-magic==0.3.0
Flask==2.0.3
future==0.18.3
gpg==1.13.1
html5lib==0.999999999
idna==2.5
importlib-metadata==4.8.3
initial-setup==0.3.81.7
isc==2.0
itsdangerous==2.0.1
Jinja2==3.0.3
langtable==0.0.51
libcomps==0.1.18
lxml==4.2.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.33
nftables==0.1
ntplib==0.3.3
ordered-set==2.0.2
perf==0.1
pexpect==4.3.1
pid==2.1.1
ply==3.9
productmd==1.11
protobuf==3.19.6
psutil==5.4.3
ptyprocess==0.5.2
pwquality==1.4.4
pycairo==1.16.3
pycups==1.9.72
pycurl==7.43.0.2
pydbus==0.6.0
pyenchant==2.0.0
pygobject==3.28.3
pykickstart==3.16.15
pymssql==0.0.0
PyMySQL==1.0.2
pyparted==3.11.7
PySocks==1.6.8
python-dateutil==2.6.1
python-linux-procfs==0.7.0
python-meh==0.47.2
pytz==2017.2
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
requests==2.20.0
requests-file==1.4.3
requests-ftp==0.3.1
rpm==4.14.3
selinux==2.9
sepolicy==1.1
setools==4.3.0
setroubleshoot==1.1
simpleline==1.1.1
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.5.1
SSSDConfig==2.7.3
syspurpose==1.28.32
systemd-python==234
typing-extensions==4.1.1
urllib3==1.24.2
virtualenv==15.1.0
webencodings==0.5.1
Werkzeug==2.0.3
zipp==3.6.0
